Fix a bug in VGR calculation
The impact of this bug was limited:
because the VGR was later log-transformed, a forgotten root caused
the data to be twice as large.
However, this did not alter the distribution of the log-transformed
data and hence did not affect the matching.
The point estimates are affected, but the p value is not.