Export a minimal anonymised dataset
We cannot export the full dataset, because this contains all data
from the anticoagulation clinics over multiple years and hence is
impossible to truly anonymise.
Instead, we export the dataset after matching.
It contains anonymised data about quality of anticoagulation,
the propensity score used for matching, and the set in which a subject
was placed based on the propensity score matching.

These data can be used to replicate the findings from
*scripts/analyse_qoac_matches.R* onwards.
Some simple adaptations are required in the script to make sure the
correct data are loaded.

In the exported data, tbi stands for time between INR measurements